Refactor: service 패키지 생성 후 위치 이동
