+++ src/main/java/english/chatbot/application/service/ChatService.java
@@ -1,4 +1,4 @@
-package english.chatbot.application;
+package english.chatbot.application.service;
 
 import english.chatbot.application.port.in.ChatUseCase;
 import jakarta.transaction.Transactional;
